+++ RapidWebView/src/main/java/in/workindia/rapidwebview/events/DownloadCompletionEvent.kt
@@ -1,6 +1,7 @@
 package `in`.workindia.rapidwebview.events
 
 import android.app.DownloadManager
+import android.content.Context
 import android.content.Intent
 import `in`.workindia.rapidwebview.constants.BroadcastConstants.Companion.EVENT_DOWNLOAD_LISTENER
 import `in`.workindia.rapidwebview.constants.BroadcastConstants.Companion.EVENT_KEY
@@ -11,11 +12,24 @@ import `in`.workindia.rapidwebview.constants.BroadcastConstants.Companion.STATUS
 import `in`.workindia.rapidwebview.constants.BroadcastConstants.Companion.SUCCESS
 import `in`.workindia.rapidwebview.broadcast.BroadcastActionHandler.generateJavaScriptEvent
 
-class DownloadCompletionEvent(private val intent: Intent) : CustomJSEvent {
+class DownloadCompletionEvent(private val context: Context, private val intent: Intent) : CustomJSEvent {
 
     override fun generateJavaScript(): String {
         val downloadId = intent.getLongExtra(DownloadManager.EXTRA_DOWNLOAD_ID, -1L)
-        val status = if (downloadId != -1L) SUCCESS else FAILURE
+        var status = if (downloadId != -1L) SUCCESS else FAILURE
+
+        val downloadManager = context.getSystemService(Context.DOWNLOAD_SERVICE) as DownloadManager
+        val query = DownloadManager.Query().setFilterById(downloadId)
+        val cursor = downloadManager.query(query)
+
+        if (cursor.moveToFirst()) {
+            val columnIndex = cursor.getColumnIndex(DownloadManager.COLUMN_STATUS)
+            if (columnIndex != -1) {
+                val statusIndex = cursor.getInt(columnIndex)
+                status = if (statusIndex == DownloadManager.STATUS_SUCCESSFUL) SUCCESS else FAILURE
+            }
+        }
+        cursor.close()
 
         return generateJavaScriptEvent(
             EVENT_DOWNLOAD_LISTENER, mapOf(